api/requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
psycopg2-binary==2.8.6
Flask-Cors==3.0.10
selenium
beautifulsoup4
python-dotenv
requests
undetected-chromedriver
numpy
pandas
regex
lxml
prettify
htmltext
